Use floats for aura values

+++ src/main/kotlin/com/github/hotm/world/auranet/AuraNode.kt
@@ -151,7 +151,7 @@ interface AuraNode {
 
     val dimPos: DimBlockPos
 
-    fun getValue(): Int
+    fun getValue(): Float
 
     fun writeToPacket(buf: NetByteBuf, ctx: IMsgWriteCtx)
